pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "aiobs"
version = "0.2.0"
description = "Minimal, extensible LLM observability SDK with OpenTelemetry, OpenAI and Gemini support."
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "aiobs Authors" }
]
keywords = ["llm", "observability", "tracing", "openai", "gemini", "genai", "opentelemetry", "otel"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries",
]

dependencies = [
  "pydantic>=2",
  "opentelemetry-api>=1.20.0",
  "opentelemetry-sdk>=1.20.0",
]

[project.optional-dependencies]
openai = [
  "openai>=1.0.0,<2",
  "opentelemetry-instrumentation-openai-v2>=2.0b0",
]
gemini = [
  "google-genai>=1.0.0",
  "opentelemetry-instrumentation-google-genai>=0.4b0",
]
vertexai = [
  "opentelemetry-instrumentation-vertexai>=2.0b0",
]
sql = [
  "sqlglot>=20.0.0",
]
example = [
  "python-dotenv>=1.0",
]
test = [
  "pytest>=7",
]
docs = [
  "sphinx>=7",
  "furo>=2024.1.29",
  "sphinx-copybutton>=0.5",
]
all = [
  "aiobs[openai,gemini,vertexai,example]",
]

[project.urls]
Homepage = "https://github.com/neuralis-in/aiobs"
Repository = "https://github.com/neuralis-in/aiobs"
Documentation = "https://neuralis-in.github.io/aiobs/"

[tool.hatch.build]
exclude = [
  ".venv",
  "example",
  ".git",
  "dist",
  "build",
]
